TP-51 Update Error Handling On Multer

+++ routes/fileUpload.js
@@ -3,37 +3,55 @@ const fs = require('fs');
 const cloudinary = require('../uploads/cloudinary');
 const upload = require('../uploads/multer');
 
+const uploadMultiple = upload.array('image', 5);
+
 const router = express.Router();
 
-router.post('/events', upload.array('image', 5), async (req, res) => {
-  try {
-    const uploader = async (path) => {
-      const result = await cloudinary.uploadsImages(path, 'Images');
-      // eslint-disable-next-line no-console
-      console.log(result);
-    };
-    const urls = [];
-    const { files } = req;
+router.post('/events', async (req, res) => {
+  uploadMultiple(req, res, async (err) => {
+    if (err) {
+      res.status(500).json({
+        message: {
+          msgBody: err.message,
+          msgError: true,
+        },
+      });
+    } else {
+      let result;
+      const uploader = async (path) => {
+        result = await cloudinary.uploadsImages(path, 'Images');
+        return result;
+      };
+      const urls = [];
+      const { files } = req;
 
-    // eslint-disable-next-line no-restricted-syntax
-    for (const file of files) {
-      const { path } = file;
-      // eslint-disable-next-line no-await-in-loop
-      const newPath = await uploader(path);
-      urls.push(newPath);
-      fs.unlinkSync(path);
+      if (files.lenght > 5 || files.lenght < 2) {
+        res.status(200).json({
+          message: {
+            msgBody: 'Should Not Above 5 Files Or Less Than 3 Files',
+            msgError: true,
+          },
+        });
+      } else {
+        // eslint-disable-next-line no-restricted-syntax
+        for (const file of files) {
+          const { path } = file;
+          // eslint-disable-next-line no-await-in-loop
+          const newPath = await uploader(path);
+          urls.push(newPath);
+          fs.unlinkSync(path);
+        }
+
+        res.status(200).json({
+          message: {
+            msgBody: 'Images Uploaded Successfully!',
+            msgError: false,
+          },
+          data: urls,
+        });
+      }
     }
-
-    res.status(200).json({
-      message: { msgBody: 'Images Uploaded Successfully!', msgError: false },
-      data: urls,
-    });
-  } catch (e) {
-    res.status(200).json({
-      message: { msgBody: 'Images Uploaded Failed!', msgError: true },
-      trace: e,
-    });
-  }
+  });
 });
 
 module.exports = router;
